Certifacte Generation/requirements.txt
# Certificate Generation System - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Dependencies
flask==2.3.3
pillow==10.0.1
qrcode[pil]==7.4.2
reportlab==4.0.4

# Database
sqlite3  # Built-in with Python

# Web Framework
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Image Processing
Pillow==10.0.1

# PDF Generation
reportlab==4.0.4

# QR Code Generation
qrcode==7.4.2

# Optional: For better performance
gunicorn==21.2.0  # For production deployment

# Development Dependencies (optional)
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0

# Security (optional)
cryptography==41.0.7
# Optional: Twilio for WhatsApp/SMS integration (install if you plan to use WhatsApp sending)
twilio==8.4.1